`tools/importer-rest-api-specs`: moving the legacy `helpers` package into the Schema package, since it's schema related
