+++ tools/importer-rest-api-specs/components/terraform/schema/helpers/snake_case.go
